Aportación funcionando

+++ fap/app/controllers/AportacionController.java
@@ -37,11 +37,13 @@ public class AportacionController extends AportacionControllerGen {
 		if (!permiso(accion)){
 			Messages.fatal("No tiene permisos suficientes para realizar esta acción");
 		}
-		Aportacion aportacion = solicitud.aportaciones.actual;
-		if(StringUtils.in(aportacion.estado, "borrador", "firmada", "registrada", "clasificada")){
-			Messages.warning("Tiene una aportación pendiente de registro");
-			Messages.keep();
-			redirect("AportacionPresentarController.index", accion, idSolicitud);
+		if(solicitud != null){
+    		Aportacion aportacion = solicitud.aportaciones.actual;
+    		if(StringUtils.in(aportacion.estado, "borrador", "firmada", "registrada", "clasificada")){
+    			Messages.warning("Tiene una aportación pendiente de registro");
+    			Messages.keep();
+    			redirect("AportacionPresentarController.index", accion, idSolicitud);
+    		}
 		}
 		renderTemplate("gen/Aportacion/Aportacion.html", accion, idSolicitud, solicitud);
 	}
@@ -66,22 +68,21 @@ public class AportacionController extends AportacionControllerGen {
 					try {
 						String tipoDocumentoSolicitudAportacion = FapProperties.get("fap.aed.tiposdocumentos.aportacion.solicitud");
 						
-						//Si los documentos ya estaban creados los borra
-						Documento borradorOld = aportacion.borrador;
-						Documento oficialOld = aportacion.oficial;						
-						aportacion.borrador = null;
-						aportacion.oficial = null;
-						aportacion.save();
+	                    // Borramos los documentos que se pudieron generar en una llamada previa al metodo, para no dejar basura en la BBDD
+						if(aportacion.borrador != null){
+						    Documento borradorOld = aportacion.borrador;
+						    aportacion.oficial = null;
+						    aportacion.save();
+						    gestorDocumentalService.deleteDocumento(borradorOld);
+						}
 						
-						gestorDocumentalService.deleteDocumento(borradorOld);
-						gestorDocumentalService.deleteDocumento(oficialOld);
-						
-						// Borramos los documentos que se pudieron generar en una llamada previa al metodo, para no dejar basura en la BBDD
-						if ((borradorOld != null) && (borradorOld.delete() == null))
-							play.Logger.info("Error borrando los documento temporales generados para el borrador");
-						if ((oficialOld != null) && (oficialOld.delete() == null))
-							play.Logger.info("Error borrando los documento temporales generados para el documento oficial");
-						
+						if(aportacion.oficial != null){
+						    Documento oficialOld = aportacion.oficial;
+						    aportacion.oficial = null;
+						    aportacion.save();
+						    gestorDocumentalService.deleteDocumento(oficialOld);
+						}						
+		
 						//Genera el borrador
 						File borrador = new Report("reports/solicitudAportacion.html").header("reports/header.html").footer("reports/footer-borrador.html").renderTmpFile(solicitud);
 						aportacion.borrador = new Documento();
@@ -99,7 +100,8 @@ public class AportacionController extends AportacionControllerGen {
 						aportacion.save();
 					}catch(Exception e){
 						Messages.error("Se produjo un error generando el documento de aportación.");
-						play.Logger.error("Error al generar el documento de la aportación: " + e.getLocalizedMessage());
+						play.Logger.error(e, "Error al generar el documento de la aportación: " + e.getMessage());
+						e.printStackTrace();
 					}
 				}
 			}
